ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.0
python-dateutil==2.8.2
joblib==1.3.2
mysql-connector-python==8.2.0
